refactor: naming convention
